feat(tableGroup): tableGroup 모듈 분리
