+++ module-tablegroup/src/main/java/kitchenpos/tablegroup/domain/TableGroupUngroupedEvent.java
