+++ module-tablegroup/src/test/java/kitchenpos/TableGroupApplicationTests.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TableGroupApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+}
