+++ module-tablegroup/src/testFixtures/java/kitchenpos/tablegroup/domain/FixtureTableGroup.java
